Add files via upload
FULL CREATION OF MODEL

The first step was to analyze the data. After analyzing the data, it seems fit for our needs for our project. I have pre-processed the data and modified the features to meet the model requirements. Normal Linear Regression was not satisfying with the results so I have used RandomForestRegression with randomSeearchCV for model tuning. Storing the best parameters to be used to actually fit and store the model. Train Data of R2 score being around 0.99 with MSE score of 0.20 and test 0.97 with MSE score of 0.16.
The model requires 11 features of year, km_driven, fuel, seller_type, transmission, owner, mileage, and engine. Model tested manually and not only with test data to mimic the functionality of being deployed in a web app.