SEARCH AI DEPLOYMENT
I have deploey a mock demo of a not so good model. I am working on a final new model which I will switch up when it is done. In the meantime I will adapt the code to it can connect to the backend and test it with picture even if it will provide wrong results for the meantime.